requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-utils==0.7.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
passlib==1.7.4
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
